+++ packages/vuikit/src/table/__dev__/table-alignment.vue
@@ -6,9 +6,9 @@
     <vk-table
       :data="data"
     >
-      <vk-table-column title="Left" cell="name"/>
-      <vk-table-column title="Center" cell="name" align="center"/>
-      <vk-table-column title="Right" cell="name" align="right"/>
+      <vk-table-column head="Left" cell="name"/>
+      <vk-table-column head="Center" cell="name" align="center"/>
+      <vk-table-column head="Right" cell="name" align="right"/>
     </vk-table>
 
     <h2>Middle</h2>
@@ -16,15 +16,15 @@
     <vk-table
       :data="[{}]"
     >
-      <vk-table-column size="small" title="Middle" align="middle">
+      <vk-table-column size="small" head="Middle" align="middle">
         <div>TABLE DATA</div>
       </vk-table-column>
-      <vk-table-column title="TABLE HEAD">
+      <vk-table-column head="TABLE HEAD">
         <div>
           Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua.
         </div>
       </vk-table-column>
-      <vk-table-column title="TABLE HEAD">
+      <vk-table-column head="TABLE HEAD">
         <vk-button size="small">BUTTON</vk-button>
       </vk-table-column>
     </vk-table>
@@ -34,15 +34,15 @@
       :data="[{}]"
       middle-aligned
     >
-      <vk-table-column size="small" title="Middle">
+      <vk-table-column size="small" head="Middle">
         <div>TABLE DATA</div>
       </vk-table-column>
-      <vk-table-column title="Middle">
+      <vk-table-column head="Middle">
         <div>
           Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua.
         </div>
       </vk-table-column>
-      <vk-table-column title="Middle">
+      <vk-table-column head="Middle">
         <vk-button size="small">BUTTON</vk-button>
       </vk-table-column>
     </vk-table>