+++ packages/vuikit/CHANGELOG.md
@@ -7,6 +7,7 @@
  - Fixed `vk-modal-full` close button trigger listener.
  - Added `vk-table-filter` and `vk-table-pagination`, `vk-table-column-date`, `vk-table-column-text` and `vk-table-column-toggle`.
  - Refactored `vk-modal` (breaking changes).
+ - Renamed `vk-table-column` `title` prop as `head`.
 
 ## 0.8.4 (April 26, 2018)